requirements.txt
PySide6>=6.0.0
EasyOCR>=1.6.0
Pillow>=9.0.0
pytest>=7.0.0
opencv-python>=4.5.0
